+++ routes/web.php
@@ -17,11 +17,13 @@ Route::get('/', function () {
     return view('login');
 });
 
+// Util
+Route::get('files/{file_id}/{view_mode}/', [\App\Http\Controllers\UtilController::class, 'view'])->name('folder.view');
+
 // Folders
 Route::post('folder/create/', [\App\Http\Controllers\FilesController::class, 'newFolder'])->name('folder.create');
 
 // Files
-Route::get('files/{file_id}/{view_mode}/', [\App\Http\Controllers\FilesController::class, 'viewFiles'])->name('folder.view');
-Route::post('files/upload/', [\App\Http\Controllers\FilesController::class, 'upFiles'])->name('files.upload');
+Route::post('files/upload/', [\App\Http\Controllers\FilesController::class, 'newFile'])->name('files.upload');
 Route::post('files/favorite', [\App\Http\Controllers\FilesController::class, 'favoriteFile'])->name('files.favorite');
 Route::get('files/download', [\App\Http\Controllers\FilesController::class, 'downloadFile'])->name('files.download');
